Main addition is mplat, a lib and DLL to optimize re-use of Rococo as a games development platform.

+++ rococo.util/rococo.ide.stdout.cpp
@@ -31,7 +31,7 @@ namespace
 
       virtual void InitDisassembly(size_t codeId)
       {
-         printf("Disassembly for id %llu\n", codeId);
+         printf("Disassembly for id %llu\n", (unsigned long long) codeId);
       }
 
       virtual void AddSourceCode(cstr name, cstr sourceCode)
